+++ integration-tests/tests/src/utils/fetch.ts
@@ -16,6 +16,6 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-import { DefaultNetworkTransport } from "realm-network-transport";
+import { DefaultNetworkTransport } from "@realm/network-transport";
 
 export const fetch = DefaultNetworkTransport.fetch;