+++ packages/realm-web/src/tests/utils/MockNetworkTransport.ts
@@ -16,7 +16,7 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-import { NetworkTransport, Request, FetchResponse, FetchHeaders } from "realm-network-transport";
+import { NetworkTransport, Request, FetchResponse, FetchHeaders } from "@realm/network-transport";
 
 import { MongoDBRealmError } from "../..";